requirements.txt
PyYAML==5.2
pytest==5.3.2
pytest-cov==2.8.1
python-interface==1.5.3
statistics==1.0.3.5
parameterized==0.7.1
flask